Création, modification, suppression et récupération de user et exam

+++ spring_pg_starter/src/main/java/org/amu/repository/CourseRepository.java
@@ -2,5 +2,5 @@ package org.amu.repository;
 
 import org.amu.model.Course;
 
-public interface CourseRepository extends CrudRepository<Course, Integer> {
+public interface CourseRepository extends CrudRepository<Course, Long> {
 }
